Replaced OverviewCards with GridLayout
