+++ site/docs/foundations/index.mdx
@@ -7,8 +7,8 @@ sidebar:
 
 Foundations are the fundamental principles that guide the creation of a cohesive design language, enabling you to build products that are easy to use and visually appealing.
 
-<OverviewCards columns={3}>
- 
+<GridLayout columns={3}>
+
    <OverviewCard
     icon={<Image src="/img/design.svg" alt="pen icon"  />}
     title="Density"
@@ -64,4 +64,4 @@ Used well typography helps to enhance readability, establish hierarchy and set a
 
   </OverviewCard>
  
-</OverviewCards>
+</GridLayout>
