+++ site/docs/getting-started/index.mdx
@@ -7,7 +7,7 @@ sidebar:
 
 Follow our step-by-step guides to design and develop accessible, responsive, best-in-class user experiences with Salt.
 
-<OverviewCards>
+<GridLayout columns={2}>
   <OverviewCard
     icon={<Image src="/img/design.svg" alt="pen icon"  />}
     title="Designing"
@@ -32,4 +32,4 @@ Learn how to install the Salt packages, integrate them into your React
 project, add the web fonts you need and import components.
 
   </OverviewCard>
-</OverviewCards>
+</GridLayout>
