+++ site/src/components/getting-started/index.ts
@@ -1,4 +1,4 @@
 export * from "./design-cards";
 export * from "./icons-table";
 export * from "./develop-fonts-accordion";
-export * from "./overview";
+export * from "./overview-card";
