+++ site/src/components/getting-started/overview-card/OverviewCard.tsx
@@ -1,0 +1,14 @@
+import { Card, CardProps } from "../../card/Card";
+
+export const OverviewCard = ({
+  keylineColor,
+  children,
+  ...rest
+}: CardProps) => (
+  <Card
+    keyLineAnimation={false}
+    keylineColor={`var(--site-tertiary-accent-${keylineColor})`}
+    {...rest}
+    description={children}
+  />
+);
