+++ site/src/components/getting-started/overview-card/index.ts
@@ -1,0 +1,1 @@
+export * from "./OverviewCard";
