--- site/src/components/getting-started/overview/OverviewCards.module.css
@@ -1,16 +1,0 @@
-.overviewCards {
-  display: grid;
-  grid-template-columns: 1fr 1fr;
-  gap: calc(var(--salt-size-unit) * 3);
-  margin: calc(var(--salt-size-unit) * 3) 0;
-}
-
-.overviewCards3Col {
-  grid-template-columns: 1fr 1fr 1fr;
-}
-
-@media screen and (max-width: 996px) {
-  .overviewCards {
-    flex-direction: column;
-  }
-}
