--- site/src/components/getting-started/overview/index.ts
@@ -1,1 +1,0 @@
-export * from "./OverviewCards";
